feat(core): introduce `callRoute` API method on Fragments
This allows users to call routes directly instead of using services.
This is useful is server-side settings such as React Router/Remix
actions when the Fragment wants to set headers.

The `createFragmentForTest` signature has also changed to accept routes
as the second parameter, and the test fragment's `callRoute` now uses
the same signature as the main API.